refactor python
