+++ setup.cfg
@@ -16,21 +16,26 @@ classifiers =
     License :: OSI Approved :: GNU General Public License v3 (GPLv3)
     License :: OSI Approved :: GNU Lesser General Public License v3
     Programming Language :: Python :: 3
+    Programming Language :: Python :: 3 :: Only
 download_url = https://github.com/cpp-lln-lab/bidspm
 provides = bidspm
 
 [options]
+packages = find:
 install_requires =
     bsmschema
     rich
+python_requires = >=3.8
+include_package_data = True
+zip_safe = False
 
 [options.entry_points]
 console_scripts =
-    bidspm = cli:main
-    bidspm_low = cli:low
-    bidspm_preproc = cli:preproc
-    bidspm_stats = cli:stats
-    validate_model = validate:main
+    bidspm = src.cli:main
+    bidspm_low = src.cli:low
+    bidspm_preproc = src.cli:preproc
+    bidspm_stats = src.cli:stats
+    validate_model = src.validate:main
 
 [options.extras_require]
 demo =
